доработка функциональности по заданию 14 спринта

+++ controllers/cards.js
@@ -1,45 +1,53 @@
+/* eslint-disable no-else-return */
 /* eslint-disable consistent-return */
+const BadRequest = require('../errors/badrequest-error');
+const NotFoundError = require('../errors/notfound-error');
 const Card = require('../models/card');
 const httpConstants = require('../utils/constants');
 
-const getCards = (_req, res) => {
+const getCards = (_req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch(() => res.status(httpConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' }));
+    .catch((next));
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+const deleteCard = (req, res, next) => {
+  const owner = req.user._id;
+  Card.findById(req.params.cardId)
     .orFail(new Error('notValidId'))
     .then((card) => {
-      res.send(card);
+      if (card.owner.toString() !== owner) {
+        return next(new BadRequest('нет прав на удаление карточки'));
+      }
+      return Card.findByIdAndRemove(req.params.cardId);
     })
+    .then((card) => res.send(card))
     .catch((err) => {
       if (err.message === 'notValidId') {
-        res.status(httpConstants.HTTP_STATUS_NOT_FOUND).send({ message: `Карточка id: ${req.params.cardId} не найдена` });
-      } else if (err.kind === 'ObjectId') {
-        res.status(httpConstants.HTTP_STATUS_BAD_REQUEST).send({ message: `Некорректные данные: ${req.params.cardId}` });
+        return next(new NotFoundError(`Карточка id: ${req.params.cardId} не найдена`));
+      } if (err.kind === 'ObjectId') {
+        return next(new BadRequest(`Некорректные данные: ${req.params.cardId}`));
       } else {
-        res.status(httpConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        return next(err);
       }
     });
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const owner = req.user._id;
   const { name, link } = req.body;
   Card.create({ name, link, owner })
     .then((card) => res.status(httpConstants.HTTP_STATUS_CREATED).send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(httpConstants.HTTP_STATUS_BAD_REQUEST).send({ message: `Некорректные данные: ${name}, ${link}` });
+        return next(new BadRequest(`Некорректные данные: ${name}, ${link}`));
       } else {
-        res.status(httpConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        return next(err);
       }
     });
 };
 
-const delLike = (req, res) => {
+const delLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -51,16 +59,16 @@ const delLike = (req, res) => {
     })
     .catch((err) => {
       if (err.message === 'notValidId') {
-        res.status(httpConstants.HTTP_STATUS_NOT_FOUND).send({ message: `Карточка id: ${req.params.cardId} не найдена` });
+        return next(new NotFoundError(`Карточка id: ${req.params.cardId} не найдена`));
       } else if (err.kind === 'ObjectId') {
-        res.status(httpConstants.HTTP_STATUS_BAD_REQUEST).send({ message: `Некорректные данные: ${req.params.cardId}` });
+        return next(new BadRequest(`Некорректные данные: ${req.params.cardId}`));
       } else {
-        res.status(httpConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        return next(err);
       }
     });
 };
 
-const putLike = (req, res) => {
+const putLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -72,11 +80,11 @@ const putLike = (req, res) => {
     })
     .catch((err) => {
       if (err.message === 'notValidId') {
-        res.status(httpConstants.HTTP_STATUS_NOT_FOUND).send({ message: `Карточка id: ${req.params.cardId} не найдена` });
+        return next(new NotFoundError(`Карточка id: ${req.params.cardId} не найдена`));
       } else if (err.kind === 'ObjectId') {
-        res.status(httpConstants.HTTP_STATUS_BAD_REQUEST).send({ message: `Некорректные данные: ${req.params.cardId}` });
+        return next(new BadRequest(`Некорректные данные: ${req.params.cardId}`));
       } else {
-        res.status(httpConstants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        return next(err);
       }
     });
 };
